Decorator for @get_data
